pyproject.toml
[build-system]
requires = ["setuptools>=61.0", "wheel"]
build-backend = "setuptools.build_meta"

[project]
name = "kenkui"
version = "0.9.0"
description = "Convert Ebooks to Audiobooks with [custom] voice samples"
readme = "README.md"
authors = [{ name = "Sumner MacArthur", email = "spn1kolat3sla@gmail.com" }]
license = "GPL-3.0-or-later"
requires-python = ">=3.12"
classifiers = [
  "Programming Language :: Python :: 3.12",
  "Operating System :: OS Independent",
]

dependencies = [
  "beautifulsoup4>=4.14.0",
  "EbookLib>=0.20",
  "textual>=1.0.0",
  "textual-fspicker>=1.0.0",
  "scipy>=1.17.0",
  "huggingface_hub>=1.3.0",
  "pydub>=0.25.0",
  "pocket-tts>=1.0.0",
  "mutagen>=1.45.0",
  "imageio-ffmpeg>=0.5.0",
  "mobi>=0.4.0",
  "pyyaml>=6.0.0",
  "fastapi>=0.115.0",
  "uvicorn>=0.34.0",
  "httpx>=0.28.0",
]

[project.scripts]
kenkui = "kenkui.__main__:main"

[project.optional-dependencies]
dev = ["pytest>=7.0.0", "pytest-cov>=4.0.0"]

[project.urls]
"Homepage" = "https://github.com/D1zzl3D0p/kenkui"
"Bug Tracker" = "https://github.com/D1zzl3D0p/kenkui/issues"

[tool.setuptools.packages.find]
where = ["src"]

[tool.setuptools.package-data]
"kenkui.samples" = ["*.epub"]
"kenkui.voices" = ["*.wav"]

[tool.pytest.ini_options]
testpaths = ["tests"]
addopts = "--tb=short"

[tool.ruff]
line-length = 100
target-version = "py312"

[tool.ruff.lint]
select = ["E", "F", "I", "N", "UP", "B", "SIM"]
ignore = [
  "E501",   # line too long — handled by formatter / not enforced
  "E402",   # module-level import not at top — intentional in several modules
  "N806",   # variable in function should be lowercase — Windows constants
  "SIM105", # use contextlib.suppress — try/except pass is clearer at module level
  "SIM108", # use ternary operator — sometimes if/else is more readable
  "B007",   # loop variable not used — some loop patterns are intentional
]

[tool.ruff.lint.isort]
known-first-party = ["kenkui"]
